// src/modules/magic/enums/condition.enum.ts
/**
 * Enum para las condiciones de las cartas
 * NM (Near Mint): Casi perfecto, carta con poco o ningún juego
 * LP (Lightly Played): Ligeramente jugada
 * MP (Moderately Played): Moderadamente jugada
 * HP (Heavily Played): Muy jugada, con bastante desgaste
 * D (Damaged): Dañada
 */
export enum EnumCondition {
  NearMint = 'NM',        // Casi perfecto, carta con poco o ningún juego
  LightlyPlayed = 'LP',   // Ligeramente jugada
  ModeratelyPlayed = 'MP', // Moderadamente jugada
  HeavilyPlayed = 'HP',   // Muy jugada, con bastante desgaste
  Damaged ="D" , // Dañada
}